refactor(helper-agent): add base agent refactot(helper-agent): moved agent code to base "agents" directory
