--- ml/agent/schemas/rag_state.py
@@ -1,12 +1,0 @@
-from pydantic import BaseModel
-from langgraph.graph import MessagesState
-from langchain_core.documents import Document
-import typing as tp
-
-class RAGState(BaseModel):
-    uuid: str
-    assignment_id: str
-    query: str
-    docs: tp.Optional[str] = None
-    msg_state: dict
-
