+++ ml/agents/helper_agent/nodes/query_rag_llm.py
@@ -1,24 +1,25 @@
-from agent.schemas.rag_state import RAGState
+from agents.helper_agent.schemas.rag_state import RAGState
 from langchain_core.messages import HumanMessage, AIMessage
 from langgraph.graph import MessagesState
 from langchain_core.language_models import BaseLanguageModel
-from agent.utils import format_model_response, format_prompt
+from agents.helper_agent.utils import format_model_response, format_prompt
 from transformers import PreTrainedTokenizerBase
 
-def query_llm(
+
+def query_rag_llm(
         state: RAGState,
         llm: BaseLanguageModel,
         tokenizer: PreTrainedTokenizerBase
     ) -> dict:
 
     messages = state.msg_state["messages"]
-
+    
     prompt = format_prompt(
         tokenizer=tokenizer,
         user_message=state.query,
         chat_history=state.msg_state["messages"],
+        context=state.docs
     )
-
     response = llm.invoke(prompt)
 
     new_messages = messages + [
