+++ ml/agents/helper_agent/nodes/retrieve.py
@@ -1,8 +1,8 @@
-from agent.schemas.rag_state import RAGState
+from agents.helper_agent.schemas.rag_state import RAGState
 from langchain_core.retrievers import BaseRetriever
 
 
-def retrieve(state: RAGState, retriever: BaseRetriever) -> str:
+def retrieve(state: RAGState, retriever: BaseRetriever) -> dict[str, str]:
     """Retrieve relevant (< threshold) information related to a query."""
     retrieved_docs = retriever.invoke(
         state.query,
