+++ ml/agents/helper_agent/nodes/route_rag_usage.py
@@ -1,4 +1,4 @@
-from agent.schemas.rag_state import RAGState
+from agents.helper_agent.schemas.rag_state import RAGState
 
 def route_rag_usage(state: RAGState) -> str:
     return "query_rag_llm" if state.docs else "query_llm"
