+++ ml/agents/schemas/agent_state.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+import typing as tp
+
+class AgentState(BaseModel):
+    uuid: str
+    assignment_id: str
+    query: str
+    msg_state: dict
